backport rockchip,rk2928-dw-mshc dw_mmc SDMMC driver from 4.9-rc6 mainline
prior to this sdmmc, slot was non-operating with 3.10

while the backport was tested ok with no issues (using MK908iii,
which is compatible to rikomagic's mk802iv and also radxa rock),
i sincerly recommend thorough testing: use the port at your own
risk, share code if you have a fix or know better.

keeping the modularity was tried, but some dependencies, such as
set_clk_phase, in particular had to be backported as well to
satisfy functional deps.

also note that some modifications to the device tree files were
done.

note that this is the second attempt at getting sdmmc working
with this kernel:  a first attempt (forward port from 3.0.36+)
was unsuccessful, kernel compiled, but driver was not DT aware..